+++ infinite_scroll_pagination/paginator.py
@@ -2,138 +2,246 @@
 
 from __future__ import unicode_literals
 
-from django.core.paginator import EmptyPage, Page
+try:
+    from collections.abc import Sequence
+except ImportError:
+    from collections import Sequence
+
+import six
+
+from django.core.paginator import EmptyPage
+from django.db.models import QuerySet
+
+__all__ = [
+    'SeekPaginator',
+    'SeekPage',
+    'EmptyPage',
+    'NEXT_PAGE',
+    'PREV_PAGE']
 
 
-__all__ = ["SeekPaginator", "SeekPage", "EmptyPage"]
+NEXT_PAGE = 1
+PREV_PAGE = 2
+
+
+class _NoPk:
+    def __str__(self):
+        return 'NoPk'
+
+    def __repr__(self):
+        return '<NoPk>'
+
+# We need a distinct value for None,
+# because None is a valid pk when fetching
+# the first page
+_NO_PK = _NoPk()
 
 
 class SeekPaginator(object):
 
     def __init__(self, query_set, per_page, lookup_field):
+        assert isinstance(query_set, QuerySet), 'QuerySet expected'
+        assert isinstance(per_page, six.integer_types), 'Int expected'
+        assert isinstance(lookup_field, six.string_types), 'String expected'
         self.query_set = query_set
         self.per_page = per_page
-        self.lookup_field = lookup_field
+        self.is_desc = lookup_field.startswith('-')
+        self.is_asc = not self.is_desc
+        self.lookup_field = lookup_field.lstrip('-')
 
-    def prepare_order(self):
-        lookup_field_desc = "-%s" % self.lookup_field
+    def prepare_order(self, has_pk, move_to):
+        pk_sort = 'pk'
+        lookup_sort = self.lookup_field
+        if ((self.is_desc and move_to == NEXT_PAGE) or
+                (self.is_asc and move_to == PREV_PAGE)):
+            pk_sort = '-%s' % pk_sort
+            lookup_sort = '-%s' % lookup_sort
+        if has_pk:
+            return [lookup_sort, pk_sort]
+        return [lookup_sort]
 
-        if self.lookup_field not in ("pk", "id"):
-            return [lookup_field_desc, "-pk"]
-        else:
-            return [lookup_field_desc, ]
-
-    def prepare_lookup(self, value, pk):
-        """
-        Lookup:
-
-        ...
-        WHERE date <= ?
-        AND NOT (date = ? AND id >= ?)
-        ORDER BY date DESC, id DESC
-        """
-        if self.lookup_field not in ("pk", "id"):
-            lookup = "%s__lte" % self.lookup_field
-            lookup_exclude = {self.lookup_field: value, "pk__gte": pk, }
-        else:
-            lookup = "%s__lt" % self.lookup_field
-            lookup_exclude = None
-
-        lookup_filter = {lookup: value, }
+    def prepare_lookup(self, value, pk, move_to):
+        lookup_include = '%s__gt' % self.lookup_field
+        lookup_exclude_pk = 'pk__lte'
+        if ((self.is_desc and move_to == NEXT_PAGE) or
+                (self.is_asc and move_to == PREV_PAGE)):
+            lookup_include = '%s__lt' % self.lookup_field
+            lookup_exclude_pk = 'pk__gte'
+        lookup_exclude = None
+        if pk is not _NO_PK:
+            lookup_include = "%se" % lookup_include
+            lookup_exclude = {self.lookup_field: value, lookup_exclude_pk: pk}
+        lookup_filter = {lookup_include: value}
         return lookup_filter, lookup_exclude
 
-    def page(self, value=None, pk=None):
-        if (value is None and pk is not None) or (value is not None and pk is None):
-            raise ValueError("Both 'value' and 'pk' arguments must be provided")
-
+    def apply_filter(self, value, pk, move_to):
         query_set = self.query_set
+        lookup_filter, lookup_exclude = self.prepare_lookup(
+            value=value, pk=pk, move_to=move_to)
+        query_set = query_set.filter(**lookup_filter)
+        if lookup_exclude:
+            query_set = query_set.exclude(**lookup_exclude)
+        return query_set
 
-        if value is not None and pk is not None:
-            lookup_filter, lookup_exclude = self.prepare_lookup(value, pk)
-            query_set = query_set.filter(**lookup_filter)
+    def seek(self, value, pk, move_to):
+        """
+        Skip the current record in case
+        there are multiple with the same value
 
-            if lookup_exclude:
-                query_set = query_set.exclude(**lookup_exclude)
+        Lookup next page (-DESC)::
 
-        order = self.prepare_order()
-        query_set = query_set.order_by(*order)[:self.per_page + 1]
+            # ...
+            WHERE date <= ?
+            AND NOT (date = ? AND id >= ?)
+            ORDER BY date DESC, id DESC
 
-        object_list = list(query_set)
-        has_next = len(object_list) > self.per_page
-        object_list = object_list[:self.per_page]
+        """
+        query_set = self.query_set
+        if value is not None:
+            query_set = self.apply_filter(
+                value=value, pk=pk, move_to=move_to)
+        query_set = query_set.order_by(
+            *self.prepare_order(
+                has_pk=pk is not _NO_PK, move_to=move_to))
+        return query_set
 
-        if not object_list and value:
-            raise EmptyPage("That page contains no results")
+    def page(self, value, pk=_NO_PK, move_to=NEXT_PAGE):
+        """
+        The param ``value`` may be ``None`` to get the first page.
+        Pass both ``value`` and ``pk`` when the ``lookup_field``'s model
+        field is not unique. Otherwise, just pass the ``value``
 
-        return SeekPage(object_list=object_list, number=value, paginator=self, has_next=has_next)
+        :raises: ``EmptyPage``
+        """
+        assert move_to in (NEXT_PAGE, PREV_PAGE)
+        query_set = self.seek(
+            value=value, pk=pk, move_to=move_to)
+
+        if value and not query_set.exists():
+            raise EmptyPage()
+
+        return SeekPage(
+            query_set=query_set,
+            key={'value': value, 'pk': pk},
+            move_to=move_to,
+            paginator=self)
 
 
-class SeekPage(Page):
+class SeekPage(Sequence):
 
-    def __init__(self, object_list, number, paginator, has_next):
-        super(SeekPage, self).__init__(object_list, number, paginator)
-        self._has_next = has_next
-        self._objects_left = None
-        self._pages_left = None
+    def __init__(self, query_set, key, move_to, paginator):
+        self._query_set = query_set
+        self._key = key
+        self._move_to = move_to
+        self._object_list = None
+        self.paginator = paginator
 
     def __repr__(self):
-        return '<Page value %s>' % self.number or ""
+        return '<Page value={value} pk={pk}>'.format(**self._key)
+
+    def __len__(self):
+        return len(self.object_list)
+
+    def __getitem__(self, index):
+        return self.object_list[index]
+
+    @property
+    def object_list(self):
+        if self._object_list is not None:
+            return self._object_list
+        # We could fetch an extra item an
+        # save a query on has_prev/next_page,
+        # but meh, the query for hax_X should be fairly cheap
+        self._object_list = list(self._query_set[:self.paginator.per_page])
+        if self._move_to == PREV_PAGE:
+            self._object_list.reverse()
+        return self._object_list
+
+    def _some_seek(self, direction):
+        assert self.object_list
+        assert direction in (NEXT_PAGE, PREV_PAGE)
+        last = self.object_list[0]
+        if direction == NEXT_PAGE:
+            last = self.object_list[-1]
+        pk = _NO_PK
+        if self._key['pk'] is not _NO_PK:
+            pk = last.pk
+        return self.paginator.seek(
+            value=getattr(last, self.paginator.lookup_field),
+            pk=pk,
+            move_to=direction)
 
     def has_next(self):
-        return self._has_next
+        if not self.object_list:
+            return False
+        return self._some_seek(NEXT_PAGE).exists()
 
     def has_previous(self):
-        raise NotImplementedError
+        if not self.object_list:
+            return False
+        return self._some_seek(PREV_PAGE).exists()
 
-    def has_other_pages(self):
-        return self.has_next()
-
-    def next_page_number(self):
-        raise NotImplementedError
-
-    def previous_page_number(self):
-        raise NotImplementedError
-
-    def start_index(self):
-        raise NotImplementedError
-
-    def end_index(self):
-        raise NotImplementedError
-
-    @property
-    def objects_left(self):
-        """
-        Returns the total number of *objects* left
-        """
-        if not self.has_next():
+    def next_objects_left(self, limit=None):
+        """Return the number of next records"""
+        if not self.object_list:
             return 0
+        qs = self._some_seek(NEXT_PAGE)
+        if limit:
+            qs = qs[:limit]
+        return qs.count()
 
-        if self._objects_left is None:
-            last = self.object_list[-1]
-            value = getattr(last, self.paginator.lookup_field)
-            lookup_filter, lookup_exclude = self.paginator.prepare_lookup(value, last.pk)
-            query_set = self.paginator.query_set.filter(**lookup_filter)
-
-            if lookup_exclude:
-                query_set = query_set.exclude(**lookup_exclude)
-
-            order = self.paginator.prepare_order()
-            self._objects_left = query_set.order_by(*order).count()
-
-        return self._objects_left
-
-    @property
-    def pages_left(self):
-        """
-        Returns the total number of *pages* left
-        """
-        if not self.objects_left:
+    def prev_objects_left(self, limit=None):
+        """Return the number of prev records"""
+        if not self.object_list:
             return 0
+        qs = self._some_seek(PREV_PAGE)
+        if limit:
+            qs = qs[:limit]
+        return qs.count()
 
-        if self._pages_left is None:
-            self._pages_left = (-self.objects_left // self.paginator.per_page) * -1  # ceil
+    def _some_pages_left(self, direction, limit):
+        some_objects_left = self.prev_objects_left
+        if direction == NEXT_PAGE:
+            some_objects_left = self.next_objects_left
+        limit = (limit or 0) * self.paginator.per_page
+        return (-some_objects_left(limit) // self.paginator.per_page) * -1  # ceil
 
-        return self._pages_left
+    def next_pages_left(self, limit=None):
+        """Return the number of next pages"""
+        return self._some_pages_left(NEXT_PAGE, limit)
 
-    def next_page_pk(self):
-        return self.object_list[-1].pk
+    def prev_pages_left(self, limit=None):
+        """Return the number of prev pages"""
+        return self._some_pages_left(PREV_PAGE, limit)
+
+    def _some_page(self, index):
+        if not self.object_list:
+            return {}
+        key = {
+            'value': getattr(
+                self.object_list[index],
+                self.paginator.lookup_field)}
+        if self._key['pk'] is not _NO_PK:
+            key['pk'] = self.object_list[index].pk
+        return key
+
+    def next_page(self):
+        """Return ``{'value': value, 'pk' pk}`` to fetch the next page"""
+        return self._some_page(-1)
+
+    def prev_page(self):
+        """Return ``{'value': value, 'pk' pk}`` to fetch the prev page"""
+        return self._some_page(0)
+
+
+def paginate(query_set, per_page, lookup_field, value, pk=None, move_to=NEXT_PAGE):
+    """Return a ``SeekPage`` containing the paginated result"""
+    return (
+        SeekPaginator(
+            query_set=query_set,
+            per_page=per_page,
+            lookup_field=lookup_field)
+        .page(
+            value=value,
+            pk=pk,
+            move_to=move_to))