+++ tests/views.py
@@ -4,33 +4,42 @@ from __future__ import unicode_literals
 import json
 
 from django.http import Http404, HttpResponse
-from django.shortcuts import get_object_or_404
 
-from infinite_scroll_pagination.paginator import SeekPaginator, EmptyPage
+from infinite_scroll_pagination import paginator
+from infinite_scroll_pagination import serializers
 
 from .models import Article
 
 
-def pagination_ajax(request, pk=None):
+def pagination_ajax(request):
     if not request.is_ajax():
         return Http404()
 
-    if pk is not None:
-        date = get_object_or_404(Article, pk=pk).date
-    else:
-        date = None
-
-    articles = Article.objects.all()
-    paginator = SeekPaginator(articles, per_page=20, lookup_field="date")
-
     try:
-        page = paginator.page(value=date, pk=pk)
-    except EmptyPage:
+        value, pk = serializers.page_key(request.GET.get('p', ''))
+    except serializers.InvalidPage:
         return Http404()
 
-    articles_list = [{"title": a.title, } for a in page]
-    data = {'articles': articles_list,
+    try:
+        page = paginator.paginate(
+            query_set=Article.objects.all(),
+            lookup_field='-date',
+            value=value,
+            pk=pk,
+            per_page=20,
+            move_to=paginator.NEXT_PAGE)
+    except paginator.EmptyPage:
+        data = {'error': "this page is empty"}
+    else:
+        data = {
+            'articles': [{'title': article.title} for article in page],
             'has_next': page.has_next(),
-            'pk': page[-1].pk}
+            'has_prev': page.has_previous(),
+            'next_objects_left': page.next_objects_left(limit=100),
+            'prev_objects_left': page.prev_objects_left(limit=100),
+            'next_pages_left': page.next_pages_left(limit=100),
+            'prev_pages_left': page.prev_pages_left(limit=100),
+            'next_page': serializers.to_page_key(**page.next_page()),
+            'prev_page': serializers.to_page_key(**page.prev_page())}
 
     return HttpResponse(json.dumps(data), content_type="application/json")